Use datadir for erb templates.
This requires putting perquackey in the Gemfile, so that
Bundler/Rubygems can properly set up the datadir for running locally.

Since relative paths in the Gemfile don't seem to be working, I had to
call File.expand_path, which would leave system-specifc information in
Gemfile.lock. So, I whacked Gemfile.lock.

This also means that running locally with bundle exec is now required,
so there's not much point in having the rake exec task.